+++ typescript/utils/src/index.ts
@@ -1,3 +1,4 @@
 export * as types from './types';
 export * as utils from './utils';
+export * as logging from './logging';
 export { Validator, BaseValidator } from './validator';
